Use size_t when representing size of memory allocation

+++ code/qcommon/qcommon.h
@@ -633,7 +633,7 @@ void Hunk_ClearToMark(void);
 void Hunk_SetMark(void);
 bool Hunk_CheckMark(void);
 void Hunk_ClearTempMemory(void);
-void* Hunk_AllocateTempMemory(int size);
+void* Hunk_AllocateTempMemory(size_t size);
 void Hunk_FreeTempMemory(void* buf);
 int Hunk_MemoryRemaining(void);
 void Hunk_Log(void);
